chore: cache room and users between bridges
