+++ apps/meteor/app/apps/server/converters/messages.js
@@ -1,9 +1,12 @@
 import { Messages, Rooms, Users } from '@rocket.chat/models';
 import { Random } from '@rocket.chat/random';
 
+import { cachedFunction } from './cachedFunction';
 import { transformMappedData } from './transformMappedData';
 
 export class AppMessagesConverter {
+	mem = new WeakMap();
+
 	constructor(orch) {
 		this.orch = orch;
 	}
@@ -18,6 +21,15 @@ export class AppMessagesConverter {
 		if (!msgObj) {
 			return undefined;
 		}
+
+		const cache =
+			this.mem.get(msgObj) ??
+			new Map([
+				['room', cachedFunction(this.orch.getConverters().get('rooms').convertById.bind(this.orch.getConverters().get('rooms')))],
+				['user', cachedFunction(this.orch.getConverters().get('users').convertById.bind(this.orch.getConverters().get('users')))],
+			]);
+
+		this.mem.set(msgObj, cache);
 
 		const map = {
 			id: '_id',
@@ -37,7 +49,7 @@ export class AppMessagesConverter {
 			token: 'token',
 			blocks: 'blocks',
 			room: async (message) => {
-				const result = await this.orch.getConverters().get('rooms').convertById(message.rid);
+				const result = await cache.get('room')(message.rid);
 				delete message.rid;
 				return result;
 			},
@@ -49,7 +61,7 @@ export class AppMessagesConverter {
 					return undefined;
 				}
 
-				return this.orch.getConverters().get('users').convertById(editedBy._id);
+				return cache.get('user')(editedBy._id);
 			},
 			attachments: async (message) => {
 				const result = await this._convertAttachmentsToApp(message.attachments);
@@ -61,7 +73,7 @@ export class AppMessagesConverter {
 					return undefined;
 				}
 
-				let user = await this.orch.getConverters().get('users').convertById(message.u._id);
+				let user = await cache.get('user')(message.u._id);
 
 				// When the sender of the message is a Guest (livechat) and not a user
 				if (!user) {
